requirements.txt
asgiref==3.4.1
Django==3.2.5
django-stubs==1.8.0
django-stubs-ext==0.2.0
djangorestframework==3.12.4
mypy==0.910
mypy-extensions==0.4.3
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
